Enforce database creation

+++ src/scenes/studying/mod.rs
@@ -123,7 +123,7 @@ pub fn draw_studying_scene(
     let (viewport_x, viewport_y) = canvas_manager.logical_to_screen(0.0, CONTENT_TOP);
     let viewport_w = logical_width;
     let viewport_h = logical_height - CONTENT_TOP;
-    clip_cam.viewport = Some((viewport_x as i32, viewport_y as i32 - (CONTENT_TOP + 10.0) as i32 , viewport_w as i32 * scale as i32, viewport_h as i32 * scale as i32));
+    clip_cam.viewport = Some((viewport_x as i32, viewport_y as i32 - (CONTENT_TOP + 10.0) as i32 , (viewport_w * scale) as i32, (viewport_h * scale) as i32));
 
 
     // Draw progress bar
